Front End static files
Landing page for  the website. We are also using Wow Book, a jquery plugin for creating the Book of quotes. Some of the files and images might not be required (came along with wow book) but I haven't done the clean up part. There are several things to be done here.
1: Currently index (list) of quotes doesn't come on every page. We need the index button to work as index of a book. Clicking on it at any point, should show a list of all the quotes. As our book gets thicker, we would need this feature for book marking the quote and also browsing, closing the book.
2: resizing of the book. Currently the book looks a bit small on the screen.